requirements.txt
djangorestframework
Django
gunicorn
django-heroku